[PATCH] Init - make sure the last parameter to execve is NULL
We alloc exactly the number of parameters in parse_line_action.
When these parameters are for execve, which request the argv
terminated by a NULL, it may fail randomly, depends on what
is there after the end of the buffer we allocated

Extend the buffer to hold one more pointer, and make sure
it is NULL to fix this bug.

Change-Id: I180df8be3502f51f81a6abb6ebf5c156eb59c9fc
Signed-off-by: Ethan <ethan.too@gmail.com>