Add new AID shared by all apps in a profile
Automatically assigned as a GID to all apps running under the same
Android profile. Used for exposing to or sharing files with apps
running in a specific profile.

Analogous to AID_SHARED_GID only the sharing is between processes with
the same userId rather than with the same appId.

The name to be displayed for uids/gids with this AID is 'u*_everyone'.

Change-Id: I46efb749f74cb86868c5609a0ce1a6aab2f9e9c9