Do not call libcutils property_set in init through libfs_mgr
Both init and libcutils define a property_set function. The init
version sets the property directly while libcutils simply calls
__system_property_set, which sends a message to init to set the
property.

Since libfs_mgr is statically linked to libcutils, any calls to
property_set end up sending a message to init and waiting for a
response. When libfs_mgr is further statically linked to init,
this leads to init sending a message to itself when property_set
is called in fs_mgr.

Because send_prop_msg in bionic only waits for a response for
250ms, this does not cause a deadlock. However, using libcutils
to set a property in the init process is hardly a good idea.

This change removes the property_set call from fs_mgr_verity.c.
If this property is required later, it should be set elsewhere.

Change-Id: I6a28cccb1ccce960841af20a4b20c32d424b5524