Set mmap_rnd_bits to maximum value.
This is a cherry-pick of internal commit:
f6112916887eddf5c17be36a0f191de0baa2ec36 which was a revert of a
revert, specifically a  revert of commit:
c8f026fc9c80ef2ae3aa0a41f55f7670894bb675.

The above revert was meant only for AOSP, since the kernel prebuilts
were not yet available there.  The revert was reverted internally so
that internal builds, which have the appropriate kernel prebuilts,
operated properly.

The very first commit was originally done in nougat-dev and cherry-picked
to AOSP, so it was not picked up again when nougat landed in master.  Add
it now.

Bug: 27681085
Test: Builds and boots.
Change-Id: If1cb6308e61aaaabca5b5bd30df78aab49e7b0d5